requirements.txt
setuptools~=49.2.0
matplotlib~=3.2.2
numpy~=1.18.5
pandas~=1.0.5
selenium~=3.141.0
pillow~=7.2.0
pypdf2~=1.26.0
pympler~=0.9
scipy~=1.5.0
yaml~=0.2.5
pyyaml~=5.3.1
cryptography~=2.9.2
tushare~=1.2.62
sqlalchemy~=1.3.18
xlsxwriter~=1.2.9
jinja2~=2.11.2
empyrical~=0.5.5
scikit-learn~=0.23.1
requests~=2.24.0
nltk~=3.5
pytesseract~=0.3.7
opencv-python~=4.4.0.46